opt: move findConstantFilterCols to general_funcs.go
Release note: None
